Split tasks to files
