+++ src/main/java/edu/hw1/Task0.java
@@ -1,0 +1,15 @@
+package edu.hw1;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+public final class Task0 {
+    private final static Logger LOGGER = LogManager.getLogger();
+
+    private Task0() {
+    }
+
+    public static void helloWorld() {
+        LOGGER.info("Hello, world!");
+    }
+}
